use pkg-config for png/freetype by default
These packages have been installing pkg-config files for quite a while
now, so default to using them.

You can still use the old way of specifying the path to the package dir
and the code will look for a <dir>/bin/<pkg>-config script.  However,
we now assume that you're using up-to-date versions of packages so we
drop all the ugly build time logic to support the various versions.