Higher precision in scaling.
Scaling (i.e. _gdScaleOneAxis()) used to accumulate the weighted
samples in unsigned char variables.  This truncated the floating-point
result toward zero, leading to a darker image and quantization of
colour gradients.

This change switches to keeping the values in doubles and only
coverting the final sums to unsigned char.  In addition to producing a
better result, it is also about 10-15% faster.

The conversion uses a couple of casts for speed.  The behaviour is
defined by ANSI but only correct on computers using twos-complement
arithmetic.  As such, this may break on obscure microcontrollers or
1960's-era minicomputers.