Added test case to exercise all of the gdImageScale() modes.
Creates and shrinks a bunch of images using gdImageScale(), once per
mode in both palette and truecolor.  It does not use GD_DEFAULT (which
is equivalent to one of the others) or GD_WEIGHTED4 (unsupported for
scaling).

Note that this test does not check if the resulting image looks anything
like the source.