Minor enhancements to the MinGW+Msys build scripts.
Makefile now creates a zipfile containing the DLLs.

Resulting DLL no longer contains a version number in its name.  (The
Windows convention is to ship the DLL with the program using it.
Versioning the name just makes it harder for scripts to find the DLL.)

Tweaked the README a bit.