Free of NULL, Not needed.
In file gd_png.c : 
line 296 and 410  image_data is NULL and by passing as 1st parameter to function 'gdFree' is freed. 
line 297 and 411  row_pointers is NULL and by passing as 1st parameter to function 'gdFree' is freed. 

However, this free does not have any impact and free(NULL) is treated as a no-op. So gdfree not needed. 

Thanks