bmp: use double variant of ceil func
We use ceill and ceil in this code, but it's not clear we need the long
double variant of ceill here.  The input multiply is already done with
double precision (the 0.1 literal is a double), and not all C libraries
offer long double variants.  Change to ceil and see if anyone notices.

Closes issue #123.