png: fix some memory leaks w/invalid images
The use of setjmp in this function was slightly incorrect: we store
pointers to allocated memory in stack variables, but we didn't call
setjmp again afterwards.  Optimization might place those in registers
which would clobber their values on the next longjmp.  Make sure we
call setjmp after every allocation to update the state.

As part of the cleanup, unify all the error code paths so we don't
have to worry about every if statement being fully correct.