tga: fix two memory corruption bugs #159
First, if the tga input had an ID set, we'd read this into the pointer
itself (and the memory after it in the struct/heap) instead of the memory
allocated to hold it.  This bug has been around since the tga code was
first added.  We'd basically corrupt memory all the time since most tga
files out there have an ID set up (like a datestamp).

Second, make sure we reject invalid images that would cause overflows in
the allocation path.  We tried to check for this already, but we used the
wrong type (we checked uint8_t but allocated int).

Fixes #159.