build: fix GDLIB_REVISION collision
We are using GDLIB_REVISION to refer to the gd version string (the "z"
in "x.y.z"), and we are using it to control the libtool revision field.
This leads to problems where the version increases (e.g. "2.1.1") but
the libtool revision doesn't (e.g. "0").  So scripts end up seeing a
revision of "0" in their output instead of "1".

Namespace the libtool version variables with "_LT_" to avoid any more
collisions.

Fixes #140.