Fix #167: Styled thick horizontal lines are scrambled
Thick lines are drawn by gdImageFilledRectangle(), which iterates over
the x ordinate first (Z order) to apply the style pattern. While this works
fine for vertical and diagonal lines, it obviously fails for horizontal
lines, which have to be iterated over in N order.

To fix this bug, we introduce the helper _gdImageFilledRectangle() to
avoid having to add another parameter to gdImageFilledRectangle() or
to add another member to the gdImageStruct (both might break ABI
compatibility).