tests: drop test_config.h
This header provides a single define: GDTEST_TOP_DIR.  Move it to the
build CPPFLAGS instead since it's a path var and that's the recommended
way of handling these (by autotools standards), and it simplifies the
build overall.  It also harmonizes the autotool & cmake build paths.

We localize this define to the gdtest/ subdir as we don't want any of
the other tests using the path.  We have gdtest helpers instead.