Resolve #282: COPYING vs. docs/naturaldocs/license.txt
Basically, docs/naturaldocs/license.txt is a duplicate of COPYING, which
is going to be used only by Natural Docs. We remove the file and generate
it on the fly instead; also we adjust COPYING to make it easier to obtain
a pleasing result without sacrificing COPYING's readability.

The diff is best viewed with whitespace changes ignored.