Fix build failure for --without-fontconfig
If fontconfig support is disabled, the static functions font_pattern() and
useFontConfig() are never used. This can lead to build errors, and does so
with the current default settings `-Wall -Werror`. Therefore we ensure that
these functions are not compiled when they are not needed.

(cherry picked from commit 59d077aeabb8b8bde7164b84c1739c6b2c26b055)