Add a basic test for gdImageCreateFromWebp()
We're testing a lossless and a lossy Webp with alpha channel. The images
have been taken from <https://developers.google.com/speed/webp/gallery2>,
have been authored by Fizyplankton and are in the public domain.

The lossy test might fail for another libwebp version, so it appears to
be best to not cherry-pick into GD-2.2 yet.