Enable test case on Windows
This test case has been disabled for Windows, because even MSVC 14
doesn't know ftello(), which is a Posix extension to the ISO standard.

Actually, ftell() is supposed to work fine for this test too, as we
don't have to deal with a large file, and we're comparing to a fixed
expected length, anyway.