bmp: check smallest negative value for undefined behavior
oss-fuzz pointed out:
gd_bmp.c:641:18: runtime error: negation of -2147483648 cannot be represented in type 'int';
                                cast to an unsigned type to negate this value to itself

This is a bit of a false positive issue as -2147483648 is -2147483648
with gcc/clang which we check for later on.  But lets check for it up
front to avoid the undefined behavior.

(cherry picked from commit 14ca9e4b235e131710bb0b84a26a25cb767af363)