gdft: change strncpy usage to memcpy
All the uses of strncpy in here are based on strlen of the input, so
there's no need to run through an str-based func again.  Switch to a
straight memcpy.  Plus this avoids static checkers that blindly choke
on strncpy.  The code was already adding a trailing NUL byte, so that
isn't problematic either.