Fix php/php-src#10614: Consistently set saveAlphaFlag
The non-square degree rotation cases, and the 0 degree rotation case all
set saveAlphaFlag to 1. But the other special cases for 90, 180, or 270
degrees don't. This results in losing the alpha channel, and in
particular if an image is black on a transparent background, in a
full-black image. Fix it by setting the saveAlphaFlag.