Support HEIC image encoding with libkvazaar
libheif currently supports x265 and kvazaar for encoding HEIC images;
while the former is recommended, the latter has a more liberal license,
and as such is an interesting alternative.  However, kvazaar apparently
does not yet support changing the chroma subsampling, and as such,
writing HEIC images fails.  We apply basically the same fix as for
libheif < 1.9.0, where we just ignore the given chroma subsampling, and
use whatever the encoder deems suitable (likely 420).

[1] <https://github.com/strukturag/libheif?tab=readme-ov-file#compiling>