Join build and tests workflows
As is, we have four individual workflows running for each push or pull
request; while their results are gathered in the overview, they still
have individual run IDs, and so navigating the results can be tedious.
Maybe even worse, if jobs fail in multiple of these workflows, there
are multiple email notifications about that, what is annoying.

Thus, we join the workflows, but still retain the individual workflow
files.