Transparent support for writing 1bpp and 4bpp BMPs
As is, we write palette images as 8bpp BMP images.  Besides that this
is wasteful, some readers may not even be able to read 8bpp BMP images.
Thus, we write palette images with at most two colors as 1bpp BMPs,
unless compression is requested, because BMPs do not support compressed
encoding of 1bpp images; we write palette images with at most 16 colors
as 4bpp BMPs, and apply `BMP_BI_RLE4` compression if requested.

We piggy-back on the existing implementation for `BMP_BI_RLE8`; that
keeps it simple, and although it might not yield the best compression,
it ought to be good enough.

References:
<https://en.wikipedia.org/wiki/BMP_file_format>
<https://learn.microsoft.com/en-us/windows/win32/gdi/bitmap-compression>

Closes #911.