Drop useless remapping of palette in gdImageNeuQuant()
This code has been taken from pngnq.c[1] where it makes much sense,
because the image is always output in PNG format.  For our case, it
does not make sense, since the output format may be different, and we
already have respective remapping code in gd_png.c.  Furthermore, this
remapping may destroy the advantageous topological properties of the
NeuQuant color quantization[2].

And actually, that code wouldn't work as it's supposed to, since the
alpha channel cannot be 255 for libgd, and as such the remapping just
inverses the order of the palette entries.

[1] <https://sourceforge.net/projects/pngnq/>
[2] <https://web.archive.org/web/20070416070905/http://members.ozemail.com.au/~dekker/NeuQuant.pdf>, section 6