Introduce gdLayerDiff() and gdEffectDiff
`gdLayerDiff()` calculates the difference of two given colors; the
difference of each RGBA channel is assigned to the result.  The
operation is commutative.

The difference of any color `c` and `0x00000000` is always `c` (no-op);
the difference of `c` and `0x7fffffff` is the inverted color (negation).
Neither of these are particularly relevant (note that the latter is
similar to `gdImageNegate()` but also inverts the alpha channel).
However, some interesting effects can be created by only setting some
channels, and possibly not to the maximum values.

As usual, `gdEffectDiff` can be set with `gdImageAlphaBlending()` so it
is applied for each `gdImageSetPixel()` call.  A particular interesting
use case is to use `imagecopy()` to create a diff of two images.  Black
pixels indicate no difference; the "brighter" the pixels, the larger
the difference.  The result is similar to `gdTestImageDiff()` in our
test suite (although less sophisticated).