Fix fix for #320
At least `gdImageRotateNearestNeighbour()` requires a truecolor image,
so we should actually pass the truecolor clone.  To be safe regarding
future changes, we also do this for `gdImageRotateGeneric()`.  And if
rotation of zero degrees is requested, we return the already cloned
image, instead of creating another clone.

We also extend the existing test case to cover all relevant code paths.