Fix #867: Support libavif 1.0 in CMake build
libavif applies the `SameMajorVersion` policy, so libavif 1.x.y is not
considered to be compatible with 0.x.y.  To still be able to require at
least libavif 0.8.2, we first look for any libavif 1.x.y, and only if
that could not be found, we apply the old `find_package()` command.